Import Clique r143
This was the final version of Clique for Wrath of the Lich King,
before the r146 Cataclysm rewrite began.

The main reason for forking this project is that there's no easy
way to direct TinyBook users to a working, non-bugged version
of Clique for TBC, since the official CurseForge page has deleted
version r102, which was pretty decent (at least after TinyBook's
injected code fixes). But there needed to be a single, all-in-one,
working version of Clique for TBC. So this is it.

License included with the original project permits forking and
modifying the code. We'll be modifying APIs to make it compatible
with 2.4.3 again, as well as doing a ton of bugfixing/enhancing.

Quoted from the original project license:

"Redistribution and use in source and binary forms, with or without
modification, are permitted..."

(And if Cladhaire sees this project, he is welcome to import any
of my bugfixes and enhancements into the latest Clique code.)