Enhance: Added "Unitframe Tooltips" to Options
Adds a "Show your active bindings in unitframe tooltips" checkbox to the
Clique Options window, so that the user doesn't have to manually type
"/clique tooltip".

This change is intended to highlight the fact that the tooltips were
completely rewritten and redesigned from scratch, and are now extremely
clean, useful and bug-free at last (as described in the "Major rewrite
of binding tooltips" commit). They are now a first-class feature of
Clique, and users are encouraged to try them out!