Enhance: Add close-button to icon select window
Clique didn't have ANY way to "stop changing icon". So if you ever
opened the icon select window, you were FORCED to click an icon (and
change your binding's icon) to get rid of the window, which was
ridiculous.

This update adds a normal "close" button in the top-right corner, to
let people abort the icon-changing window.