Bugfix: Correctly set the target unit for actions
The old code was completely bugged. It wasn't setting the "target unit"
attribute of your click-actions if the action itself didn't have a target
unit specified.

What that meant is that if there was a lingering "target" specifier from
an old action, then your NEW action would WRONGLY target THAT old unit.

NOTE: A very easy way to trigger the bug would be to set up an action such
as "Set Focus", and set its target to "player", so that clicking ANY frame
would always target player. Then edit your action and remove the
"player" target (make it an empty string again). Clicking on any frame
would STILL make "player" the focus target. The correct result should be
that the clicked frame becomes the focus, but due to the lingering
attribute, that was completely broken...

* We now ALWAYS set ALL attributes for a given action, to ensure that we
  overwrite any lingering, old attributes. This fixes all broken
  "targeted" actions.

* This commit also cleans up the SetAttribute/DeleteAttribute code by
  removing unused variables and restructuring the code a bit, as well
  as commenting out a useless "assert()" statement.