Bugfix: Correctly toggle Delete/Edit/Max buttons
The Delete/Edit/Max bottom-bar buttons are supposed to light up
appropriately based on your binding-list selection. However,
due to a silly bug in the order of the code, it was applying
the button toggling BEFORE re-sorting the list elements.

This meant that if you added or deleted a binding, and that caused
your "selection" to move to another list entry, then the bottom-bar
buttons did NOT properly update to reflect the NEW selection.

So, for example, if you select a rankless/max-rank spell, which means
that the "Max" button is disabled, and then you ADD/DELETE some
binding which causes your SELECTION to land on a binding WITH a
rank, then the "Max" button did NOT get properly enabled, and vice
versa...

The fix was simple: FIRST re-sort the list's latest contents, THEN
toggle the bottom-bar buttons based on the RESULTING list selection.

(The old code was doing that completely backwards. Mind-boggling.)