Bugfix: Properly display invalid custom actions
When the user types unexpected data into the Custom Action window, such
as typing a word inside of a "Bag Number" field, it would break all of
Clique permanently! The whole list display of all bound actions would
actually become completely corrupt and crash the addon, so that the user
couldn't even edit their custom action to fix it!

That has now been solved, by removing type-restrictions on "argument
formatting", so that we don't DEMAND that certain arguments are numbers,
etc. Instead, everything is rendered/formatted as strings, which means
that the "corrupted binding list" situation can't happen anymore.