Bugfix: Refuse to bind passive spells in Custom
When the user creates a Custom Action, they were previously able to type
names such as "Frost Resistance" (which is a Racial Passive), and the
action would happily add itself to the list.

That's bad, since passive spells CANNOT BE CAST, so it makes zero sense
to bind them. That's why the main binding method (clicking spells in the
spellbook) refuses to add passives!

This situation has now been fixed. The Custom Action window will now
refuse to save any "Cast Spell" binding that refers to a Passive or
Racial Passive spell.