Enhance: Automatically assign Custom Action icons
Whenever you bind a spell directly in Clique by clicking on a spellbook
entry, you always get a very nice and completely accurate spell icon.

However, adding "Custom" actions was nowhere near as polished. The icon
was ALWAYS an ugly "?" (question mark), and it was tedious to change the
icon. Especially if you wanted the action's icon to reflect something
like the actual item icon for a "Use Item" action. That WASN'T EVEN
POSSIBLE, since the manual icon selector doesn't contain item icons!

But that's all in the past now! This patch adds AUTOMATIC icon
assignments for ALL custom action types, to ensure that you can always
get a quick "at a glance" view of your actions and what they do!

We now automatically scan the action (when you save it) and all of its
parameters to determine the most appropriate icon for action type.

The scanning is done as follows:

* If the user manually sets a custom icon, EXCEPT for the "?" (question
  mark) icon, then we use their manually assigned icon.

* If the user keeps the default "?" (question mark) icon, or sets
  the icon to the question mark again, then we auto-detect the icon.

- "Change ActionBar": An "eye" icon, meant to indicate the fact that
  you're changing your "action bar view".

- "Action Button": A red and white "swirl" icon, since it's absolutely
  impossible to know what's going to sit on a specific action bar button
  at any given time. So this non-descript icon is intentional.

- "Pet Action Button": The famous "sad baby seal face" pet icon.

- "Cast Spell": A case-insensitive scan is performed in the player's
  spellbook and pet book, and if the spell is found, we use the highest
  rank's icon for that spell. If nothing is found, we keep the "?".

- "Use Item": If the player links the item by shift-clicking on an item
  to create an item-link in the "Item Name" field, then we accurately
  extract the exact item icon. However, if the player MANUALLY TYPES the
  item name into the field, then we perform a case-insensitive scan in
  the client's item cache (all items seen in bag, bank, mail, etc) to
  attempt to find the icon. If nothing is found, we keep the "?".

- "Run Custom Macro": A dark-blue, tranquil night sky. Just something
  to indicate the "endless possibilities of macros". Just kidding;
  actually, there just isn't any truly appropriate "macro" icon in the
  game. But the night sky is very distinct from normal spell icons, so
  your macros will stand out in the list.

- "Stop Casting": A red "minus sign" on a red background.

- "Target Unit": A green minotaur head with a red "crosshair" on it.

- "Set Focus": A black silhouette against a blue background, with a
  white, glowing dot in the person's "chest".

- "Assist Unit": The silhouette of two hands reaching up against
  a yellow, "glowing" starburst background, meant to indicate "helping"
  someone.

- "Click Button": A tiny, blue crystal shard radiating blue pulses,
  which kind of looks like a "highlighted button".

- "Show Unit Menu": A white, glowing orb with blue outline.

With these new icons, you'll never be faced with a horrible list of ugly
"?" question mark icons again.

TIP: If you have old bindings that still use the old "?" icon, simply
edit them and save them again to apply the auto-assigned icons!