Bugfix: Auto-close the Custom Action Icon selector
When you closed the Custom Action editor (the "Custom" / "Edit" window),
it didn't properly close its Icon Selector frame, which meant that the
next time you opened the Custom editor, it still had a completely
unwanted and unrelated "Icon Selector" floating on top of it, if you
hadn't manually closed the selector BEFORE closing the Custom window.

We now AUTOMATICALLY close any lingering Icon Selector window when the
user presses either "Cancel" or "Save" on the Custom Action editor.

Note that we do NOT auto-close it "OnHide", since the Custom editor
can be hidden by all kinds of things that don't mean that the user
wanted to end their editing. For example, closing and re-opening
the spellbook would temporarily hide and then re-appear Clique!