Bugfix: Use IsShown when toggling visibility
The old code used "IsVisible", which is useful for some things, but
demands that the entire parent frame hierarchy is visible too.

Basically: IsVisible = Currently visible on screen. IsShown = Has
been marked as visible, regardless of whether it's being shown or
is invisible (due to its parents being hidden).

When toggling, we want to use "IsShown" to ensure that we properly
toggle frame states even when their parents are temporarily hidden.