Bugfix: Refuse to toggle Clique in combat
There was already an "OnShow: if in combat, immediately hide again"
handler, and a "when entering combat, hide Clique" event handler.

But this commit adds some code which gets deeper into the root of the
problem, by simply making "Clique:Toggle()" refuse to do anything
if the player is currently in combat!

All of these "hide in combat" tricks are necessary because Clique
isn't allowed to modify game bindings while the player is in combat,
so it makes no sense to even show the Clique frame at all.

This commit also cleans up some if/elseif/else nesting a bit...